# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Профессия 2.0 API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/courses": {
      "get": {
        "summary": "List courses",
        "parameters": [
          {
            "name": "skill",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/courses/{id}": {
      "get": {
        "summary": "Get course",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/courses/{id}/lessons": {
      "get": {
        "summary": "Course lessons",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/jobs": {
      "get": {
        "summary": "List jobs",
        "parameters": [
          { "name": "city", "in": "query", "schema": { "type": "string" } },
          { "name": "skills", "in": "query", "schema": { "type": "string" } },
          { "name": "matchFor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "post": { "summary": "Create job", "responses": { "201": { "description": "Created" } } }
    },
...
  },
...
}